requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
schedule>=1.2.1
requests>=2.31.0
ta>=0.11.0
lxml>=5.0.0
html5lib>=1.1
beautifulsoup4>=4.12.0
finnhub-python>=2.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
